Protectd route

+++ blog_api/resources/item.py
@@ -1,26 +1,30 @@
-import uuid
-from flask import request
-from sqlalchemy.exc import SQLAlchemyError
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from schemas import ItemSchema, ItemUpdateSchema
+from flask_jwt_extended import jwt_required
+from sqlalchemy.exc import SQLAlchemyError
+from flask_jwt_extended import jwt_required, get_jwt
+
 from db import db
 from models import ItemModel
-
-
-from db import items
+from schemas import ItemSchema, ItemUpdateSchema
 
 blp = Blueprint("Items", __name__, description="Operations on items")
 
 
 @blp.route("/item/<string:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
-        return ItemModel.query.all()
+        item = ItemModel.query.get_or_404(item_id)
+        return item
 
-
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+            
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -29,12 +33,13 @@ class Item(MethodView):
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
     def put(self, item_data, item_id):
-        item = ItemModel.query.get(item_id)
+        item = ItemModel.query.get_or_404(item_id)
+
         if item:
             item.price = item_data["price"]
             item.name = item_data["name"]
         else:
-            item = ItemModel(id=item_id, **item_data)
+            item = ItemModel(**item_data)
 
         db.session.add(item)
         db.session.commit()
@@ -44,15 +49,15 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
-
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
-        return {"items": list(items.values())}
+        return ItemModel.query.all()
 
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
-        
         item = ItemModel(**item_data)
 
         try:
